refactor(dashboard): split dashboard module into smaller components and update imports
