+++ packages/main-ui/src/pages/dashboard/controller.rs
@@ -1,16 +1,16 @@
-#![allow(non_snake_case)]
+#![allow(non_snake_case, dead_code)]
 use std::time::{Duration, UNIX_EPOCH};
 
 use chrono::{self, DateTime, Local};
 use dioxus::prelude::*;
 use dioxus_logger::tracing;
+use dioxus_translate::Language;
 use models::prelude::{ListSurveyResponse, SurveyDraftStatus, UpsertSurveyDraftRequest};
 
 use models::prelude::Survey;
 
+use crate::routes::Route;
 use crate::service::prev_survey_api::PrevSurveyApi;
-
-use super::{Language, Route};
 
 // #[derive(Debug, Clone, PartialEq)]
 // pub struct Survey {
